Primarily this change moves the app from:
(a) getting data for a single year for a single county each time the user requests data
(b) pre-generating all data for all counties for all years prior to app load, and just subsetting the data to render it.

Since we're just using counties now, remove the map.

And since we're dealing with time series now, render a line graph.