Make line graph had dots for values as well.
This solves a few problems:
1. For graphs like Hays County, TX Work from Home,
where the values oscillate between NA and a number,
the line graph shows nothing. I guess the rule there
is "only show of a line if the adjacent x value exists"
(i.e. it doesn't make lines that "skip"). So now those
values appear.
2. It makes it a bit clearer (in the line graph at least)
that 2020 has no data.