Change structure of county rankings table.
It is now: Rank, <County, State>, 2019, 2021, Change, Percent Change.

I think this does a better job of giving the users the information they want
(both absolute numbers and % change). And it avoids having separate columns
for both state and county (which may or may not be good).

A nice improvement would be if the selected county was somehow highlighted
and always present when the user clicks on this tab.