Move map to quartiles instead of continuous scale
This has the effect of making high-level geographic patterns in the data
easy to spot at a glance. Individual datapoints are still accessible via
rollover. The problem with the continuous scale was that it was dominated
by outliers.

Also: read in YEAR data as a string so that matplotlib stops adding
".0" to the end of it in the x-axis.