Change algorithm for pruning counties from dataset
Dataset generation algorithm still principally works like:

for year in years
    get data on all counties

And we still have the issue where the name and number of counties varies
each year.

But now we remove a county from the dataset iff it is not present in
both 2019 and 2021.

In short, we now say that we are only interested in counties which
we can do a % change calculation on from 2019-2021 (which corresponds
to covid).