Change graph 2 from a histogram to a boxplot.
The problem with the histogram is that outliers appear as empty space. That is, the bar is so low that you cannot tell where the value is, or how many outliers there are.

The boxplot really solves that problem. The problem is that some people will not know how to interpret it.