Remove dependencies data directory had on its parent directory.
It appears that the way Python works, life is easier if a script in a subdirectory (such as data/gen_county_data.py) does not import anything from a parent directory. Note that gen_county_data.py is a script (meaning it is meant to be run on the command line), and not a module (meaning it is not meant to be imported into another script).

In practice this meant moving the get_unique_census_labels function from /backend.py to /data/census_vars.py.

This required turning the data directory into a package (i.e. adding a __init__.py file to it) so that the function could be accessed by modules in the parent directory.

This allows me to continue to run the data generation script as "python gen_county_data.py" from the command line in the directory where it exists. It appears that the old design might have required me to run the script from a directory above and as a module (i.e. with "python -m").