Rename columns
Rename "STATE_NAME" -> "State"
Rename "COUNTY_NAME" -> "County"

These columns were originally well named, in that the raw data comes in in columns STATE and COUNTY. Those columns also come in as numeric FIPS codes. So it made sense to add the suffix _NAME to them to indicate they were names. And it also made sense to follow the all caps naming convention.

But the rest of the df uses Title Case for column names. So I think it makes more sense to follow that convention here too.