Store county's full name in dataset
Currently we store the state name and the county name in different columns. This makes sense, as e.g. in the dropdown we need them as separate fields. But in various places in the code the app passes the state and county fields as parameters only to have the called function reconstruct the county's full name (e.g. "county, state"). This seems unnecessary.

This checkin adds a new "Full Name" column to the dataset. The UI reconstructs the full name, and then passes it as a parameter. This simplifies the code.